v0.91 Checked subclassing & 3D capability - worked, but problems came up. Check the previous commit for the Sxp1 experiment.
Specifically, every bot was stored in an array of class <Bot> before,
but if bots are of different classes a cell array is necessary. That
means the <Bot> constructor has to be modified respectively.
Computation time increased considerably (mostly due to strategies).
Must fix to be compatible with the general <Bot> class again.